+++ x/data/server/query_anchor_by_iri.go
@@ -3,8 +3,7 @@ package server
 import (
 	"context"
 
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
+	regenerrors "github.com/regen-network/regen-ledger/errors"
 	"github.com/regen-network/regen-ledger/types"
 	"github.com/regen-network/regen-ledger/x/data"
 )
@@ -12,22 +11,22 @@ import (
 // AnchorByIRI queries a data anchor by the IRI of the data.
 func (s serverImpl) AnchorByIRI(ctx context.Context, request *data.QueryAnchorByIRIRequest) (*data.QueryAnchorByIRIResponse, error) {
 	if len(request.Iri) == 0 {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("IRI cannot be empty")
+		return nil, regenerrors.ErrInvalidArgument.Wrap("IRI cannot be empty")
 	}
 
 	contentHash, err := data.ParseIRI(request.Iri)
 	if err != nil {
-		return nil, err
+		return nil, regenerrors.ErrInvalidArgument.Wrapf("failed to parse IRI: %s", err.Error())
 	}
 
 	dataID, err := s.stateStore.DataIDTable().GetByIri(ctx, request.Iri)
 	if err != nil {
-		return nil, sdkerrors.ErrNotFound.Wrapf("data record with IRI")
+		return nil, regenerrors.ErrNotFound.Wrapf("data record with IRI: %s", request.Iri)
 	}
 
 	anchor, err := s.stateStore.DataAnchorTable().Get(ctx, dataID.Id)
 	if err != nil {
-		return nil, err
+		return nil, regenerrors.ErrNotFound.Wrap(err.Error())
 	}
 
 	return &data.QueryAnchorByIRIResponse{