+++ x/data/server/query_attestations_by_hash_test.go
@@ -90,13 +90,13 @@ func TestQuery_AttestationsByHash(t *testing.T) {
 
 	// query attestations with empty content hash
 	_, err = s.server.AttestationsByHash(s.ctx, &data.QueryAttestationsByHashRequest{})
-	require.EqualError(t, err, "content hash cannot be empty: invalid request")
+	require.EqualError(t, err, "content hash cannot be empty: invalid argument")
 
 	// query attestations with invalid content hash
 	_, err = s.server.AttestationsByHash(s.ctx, &data.QueryAttestationsByHashRequest{
 		ContentHash: &data.ContentHash{},
 	})
-	require.EqualError(t, err, "invalid data.ContentHash: invalid type")
+	require.EqualError(t, err, "failed to convert to IRI: invalid data.ContentHash: invalid type: invalid argument")
 
 	// query attestations with content hash that has not been anchored
 	_, err = s.server.AttestationsByHash(s.ctx, &data.QueryAttestationsByHashRequest{
@@ -106,5 +106,5 @@ func TestQuery_AttestationsByHash(t *testing.T) {
 			CanonicalizationAlgorithm: data.GraphCanonicalizationAlgorithm_GRAPH_CANONICALIZATION_ALGORITHM_URDNA2015,
 		}},
 	})
-	require.EqualError(t, err, "data record with content hash: not found")
+	require.EqualError(t, err, "data record with IRI: regen:13toVfw5KEeQwbmV733E3j9HwhVCQTxB7ojFPjGdmr7HX3kuSASGXxV.rdf: not found")
 }